Support MacOS (#20)
Fix build issues to allow use with MacOS.

- Fix compiler warnings, type casting stuff.

- Require GNU sed. MacOS version silently produces wrong results as it doesn't support GNU escapes. NB. Can this be rewritten, perhaps with python script?

- Fix #include path issue. MacOS uses case-insensitive filenames and warns about conflict between "jerryscript.h" and "Jerryscript.h". Refactor to isolate jerryscript paths which aren't required by applications

+++ src/Context.cpp
@@ -11,7 +11,6 @@
  */
 
 #include "include/Jerryscript/Context.h"
-#include <jerryscript.h>
 
 namespace Jerryscript
 {
@@ -23,7 +22,7 @@ Context::Context() : Context(JERRY_GLOBAL_HEAP_SIZE * 1024)
 
 Context::Context(size_t heapSize)
 {
-	auto ctx = jerry_create_context(std::max(heapSize, 1024U), alloc, this);
+	auto ctx = jerry_create_context(std::max(heapSize, size_t(1024U)), alloc, this);
 	context.reset(reinterpret_cast<uint8_t*>(ctx));
 }
